Nuke defines kOfxParamPropInteractMinimumSize as double x 2

+++ Engine/OfxOverlayInteract.cpp
@@ -230,11 +230,11 @@ NatronOverlayInteractSupport::n_getOverlayColour(double &r,
 }
 
 void
-Natron::OfxParamOverlayInteract::getMinimumSize(int & minW,
-                                                int & minH) const
+Natron::OfxParamOverlayInteract::getMinimumSize(double & minW,
+                                                double & minH) const
 {
-    minW = _descriptor.getProperties().getIntProperty(kOfxParamPropInteractMinimumSize,0);
-    minH = _descriptor.getProperties().getIntProperty(kOfxParamPropInteractMinimumSize,1);
+    minW = _descriptor.getProperties().getDoubleProperty(kOfxParamPropInteractMinimumSize,0);
+    minH = _descriptor.getProperties().getDoubleProperty(kOfxParamPropInteractMinimumSize,1);
 }
 
 void
